Combine base into docker files
Simplifies build process and triggers. Provides a simple refresh script to
rebuild the templates based on the template.

Unfortunately docker can't copy files from outside the build directory because
it runs the build process in isolation. This means we have to keep copies of
the support files in each folder as well. :( :( :(