Refactor into phusion/baseimage
This make startup a bit easier by splitting the entrypoint from the complex setup script.

Additionally there are some additional to parallelize the file permissions(when setting the data directory permissions) as well as some other cleanups.

There was also some issues with starting up on the ubuntu images. Strangely the 16.04 image is saying its 16.10 so something might be a bit broken...

Downside the images are bigger.